+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Raise RetryableError during process if an exception occurs
 
 ## 1.0.1 2017-11-27
 - Added logging to gateway